+++ app/build.gradle
@@ -1,12 +1,11 @@
 plugins {
     id 'com.android.application'
-    //id 'com.google.android.gms.oss-licenses-plugin'
     id 'org.jetbrains.kotlin.android'
     id 'org.jetbrains.kotlin.plugin.serialization'
 }
 
 android {
-    namespace 'com.nonnonstop.tsubonofuta'
+    namespace 'io.github.chipppppppppp.binnosoko'
     compileSdk 34
 
     defaultConfig {
@@ -64,8 +63,8 @@ android {
 dependencies {
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'androidx.preference:preference:1.2.0'
-    //implementation 'com.google.android.gms:play-services-oss-licenses:17.0.1'
+    implementation 'androidx.preference:preference-ktx:1.2.1'
+    implementation 'com.google.android.gms:play-services-oss-licenses:17.0.1'
     implementation "org.jetbrains.kotlin:kotlin-reflect:1.9.23"
     implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3'
     implementation 'org.jetbrains.kotlinx:kotlinx-serialization-properties:1.6.3'
