+++ app/src/main/AndroidManifest.xml
@@ -1,30 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@android:mipmap/sym_def_app_icon"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@android:mipmap/sym_def_app_icon"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Tsubonofuta">
+        android:theme="@android:style/Theme.Material.NoActionBar">
+
         <activity
-            android:name="space.aioilight.tsubonofuta.ui.SettingsActivity"
+            android:name="io.github.chipppppppppp.binnosoko.ui.SettingsActivity"
             android:exported="true"
-            android:label="@string/title_activity_settings"
-            android:theme="@style/Theme.Tsubonofuta.NoActionBar">
+            android:label="@string/title_activity_settings">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
         <provider
-            android:name="space.aioilight.tsubonofuta.config.ConfigProvider"
-            android:authorities="space.aioilight.tsubonofuta.provider"
+            android:name="io.github.chipppppppppp.binnosoko.config.ConfigProvider"
+            android:authorities="io.github.chipppppppppp.binnosoko.provider"
             android:enabled="true"
-            android:exported="true" />
+            android:exported="true"
+            tools:ignore="ExportedContentProvider" />
 
         <meta-data
             android:name="xposedmodule"
@@ -37,7 +38,7 @@
             android:value="82" />
         <meta-data
             android:name="xposedscope"
-            android:resource="@array/apps" />
+            android:value="jp.co.airfront.android.a2chMate" />
         <meta-data
             android:name="xposedsharedprefs"
             android:value="true" />
