+++ app/src/main/assets/xposed_init
@@ -1,1 +1,1 @@
-space.aioilight.tsubonofuta.hook.ModuleMain
+io.github.chipppppppppp.binnosoko.hook.ModuleMain
