+++ app/src/main/java/io/github/chipppppppppp/binnosoko/hook/IHook.kt
@@ -1,0 +1,11 @@
+package io.github.chipppppppppp.binnosoko.hook
+
+import de.robv.android.xposed.callbacks.XC_LoadPackage
+import io.github.chipppppppppp.binnosoko.config.ConfigResolver
+
+interface IHook {
+    fun register(
+        config: ConfigResolver,
+        lpParam: XC_LoadPackage.LoadPackageParam
+    )
+}
