+++ app/src/main/res/menu/settings_activity.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-
     <item
         android:id="@+id/advanced"
         android:title="@string/settings_advanced_header" />
