--- app/src/main/res/values/arrays.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string-array name="apps">
-        <item>jp.co.airfront.android.a2chMate</item>
-    </string-array>
-</resources>
