+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Binnosoko</string>
+    <string name="app_name" translatable="false">Binnosoko</string>
     <string name="repository_url" translatable="false">https://github.com/Chipppppppppp/Binnosoko</string>
     <string name="title_activity_settings" translatable="false">Binnosoko</string>
     <string name="settings_ad_header">Advertisement</string>
@@ -8,7 +8,7 @@
     <string name="settings_client_header">Client</string>
     <string name="settings_replace_ua_title">Enable replacing User-Agent</string>
     <string name="settings_replace_ua_desc">Modify User-Agent of ChMate client.</string>
-    <string name="settings_ua_title">User-Agent</string>
+    <string name="settings_ua_title" translatable="false">User-Agent</string>
     <string name="settings_remove_api_id_title">Make MonaKey removable</string>
     <string name="settings_remove_api_id_desc">When removing your cookie, also removes MonaKey (a post key).</string>
     <string name="settings_advanced_header">Advanced setting</string>
