--- app/src/main/res/values/themes.xml
@@ -1,6 +1,0 @@
-<resources>
-
-    <style name="Theme.Tsubonofuta" parent="@style/Theme.AppCompat.Light" />
-
-    <style name="Theme.Tsubonofuta.NoActionBar" parent="@style/Theme.AppCompat.Light.NoActionBar" />
-</resources>
