--- build.gradle
@@ -1,5 +1,0 @@
-buildscript {
-    dependencies {
-        //classpath 'com.google.android.gms:oss-licenses-plugin:0.10.6'
-    }
-}
